organized files to be separated by database type
